Deploy: add global Loading & Router setting
