+++ src/app/page.tsx
@@ -1,22 +1,5 @@
-"use client";
+import Home from "@/components/home/Home";
 
-import { useEffect, useState } from "react";
-import Dashboard from "@/components/dashboard/Dashboard";
-import Landing from "@/components/landing/Landing";
-import { useUser } from "@/hooks/useUser";
-
-const Home = () => {
-  const [isLogin, setIsLogin] = useState(false);
-  const { isSuccess, isError } = useUser();
-
-  useEffect(() => {
-    if (isSuccess) {
-      setIsLogin(true);
-    } else if (isError) {
-      setIsLogin(false);
-    }
-  }, [isSuccess, isError]);
-
-  return <>{isLogin ? <Dashboard /> : <Landing />}</>;
-};
-export default Home;
+export default function Page() {
+  return <Home />;
+}
